Header Docs (Issue #13), Get rid of compiler warning (Issue #26), Bump version for pods integration (#5)

+++ IKJayma/IKJayma/IJError.m
@@ -31,7 +31,7 @@
 	[returnString appendFormat:@"Response: %@\n", self.response];
 	[returnString appendFormat:@"Response Object: %@\n", self.responseObject];
 	[returnString appendFormat:@"Internal Error: %@\n", self.internalError.localizedDescription];
-	[returnString appendFormat:@"code: %i", self.code];
+	[returnString appendFormat:@"code: %li", (long)self.code];
 	
 	return returnString;
 }
